move state-related stuff from sim to linalg.states
