+++ cirq/__init__.py
@@ -112,10 +112,13 @@ from cirq.linalg import (
     AxisAngleDecomposition,
     bidiagonalize_real_matrix_pair_with_symmetric_products,
     bidiagonalize_unitary_with_special_orthogonals,
+    bloch_vector_from_state_vector,
     block_diag,
     CONTROL_TAG,
+    density_matrix_from_state_vector,
     diagonalize_real_symmetric_and_sorted_diagonal_matrices,
     diagonalize_real_symmetric_matrix,
+    dirac_notation,
     dot,
     expand_matrix_in_orthogonal_basis,
     hilbert_schmidt_inner_product,
@@ -143,13 +146,18 @@ from cirq.linalg import (
     partial_trace,
     PAULI_BASIS,
     scatter_plot_normalized_kak_interaction_coefficients,
+    STATE_VECTOR_LIKE,
     pow_pauli_combination,
     reflection_matrix_pow,
     slice_for_qubits_equal_to,
     so4_to_magic_su2s,
     targeted_conjugate_about,
     targeted_left_multiply,
+    to_valid_density_matrix,
+    to_valid_state_vector,
     unitary_eig,
+    validate_normalized_state,
+    von_neumann_entropy,
     wavefunction_partial_trace_as_mixture,
 )
 
@@ -303,7 +311,6 @@ from cirq.optimizers import (
 )
 
 from cirq.sim import (
-    bloch_vector_from_state_vector,
     StabilizerStateChForm,
     CIRCUIT_LIKE,
     CliffordSimulator,
@@ -311,12 +318,10 @@ from cirq.sim import (
     CliffordSimulatorStepResult,
     CliffordTableau,
     CliffordTrialResult,
-    density_matrix_from_state_vector,
     DensityMatrixSimulator,
     DensityMatrixSimulatorState,
     DensityMatrixStepResult,
     DensityMatrixTrialResult,
-    dirac_notation,
     measure_density_matrix,
     measure_state_vector,
     final_density_matrix,
@@ -333,13 +338,8 @@ from cirq.sim import (
     SimulationTrialResult,
     Simulator,
     SparseSimulatorStep,
-    STATE_VECTOR_LIKE,
     StateVectorMixin,
     StepResult,
-    to_valid_density_matrix,
-    to_valid_state_vector,
-    validate_normalized_state,
-    von_neumann_entropy,
     WaveFunctionSimulatorState,
     WaveFunctionStepResult,
     WaveFunctionTrialResult,
