+++ cirq/circuits/circuit.py
@@ -1602,10 +1602,9 @@ class Circuit:
         qid_shape = self.qid_shape(qubit_order=qs)
         state_len = np.product(qid_shape, dtype=int)
 
-        from cirq import sim
-        state = sim.to_valid_state_vector(initial_state,
-                                          qid_shape=qid_shape,
-                                          dtype=dtype).reshape(qid_shape)
+        state = linalg.to_valid_state_vector(initial_state,
+                                             qid_shape=qid_shape,
+                                             dtype=dtype).reshape(qid_shape)
         result = _apply_unitary_circuit(self, state, qs, dtype)
         return result.reshape((state_len,))
