+++ cirq/linalg/__init__.py
@@ -42,8 +42,16 @@ from cirq.linalg.decompositions import (
 )
 
 from cirq.linalg.states import (
-    one_hot,
+    bloch_vector_from_state_vector,
+    density_matrix_from_state_vector,
+    dirac_notation,
     eye_tensor,
+    one_hot,
+    STATE_VECTOR_LIKE,
+    to_valid_density_matrix,
+    to_valid_state_vector,
+    validate_normalized_state,
+    von_neumann_entropy,
 )
 
 from cirq.linalg.diagonalize import (
