+++ cirq/sim/density_matrix_simulator.py
@@ -19,7 +19,7 @@ from typing import Dict, Iterator, List, TYPE_CHECKING, Type, Union
 
 import numpy as np
 
-from cirq import circuits, ops, protocols, study, value, devices
+from cirq import circuits, ops, linalg, protocols, study, value, devices
 from cirq.sim import density_matrix_utils, simulator
 
 if TYPE_CHECKING:
@@ -253,11 +253,10 @@ class DensityMatrixSimulator(simulator.SimulatesSamples,
             circuit.all_qubits())
         qid_shape = protocols.qid_shape(qubits)
         qubit_map = {q: i for i, q in enumerate(qubits)}
-        initial_matrix = density_matrix_utils.to_valid_density_matrix(
-            initial_state,
-            len(qid_shape),
-            qid_shape=qid_shape,
-            dtype=self._dtype)
+        initial_matrix = linalg.to_valid_density_matrix(initial_state,
+                                                        len(qid_shape),
+                                                        qid_shape=qid_shape,
+                                                        dtype=self._dtype)
         measured = collections.defaultdict(
             bool)  # type: Dict[Tuple[cirq.Qid, ...], bool]
         if len(circuit) == 0:
@@ -400,7 +399,7 @@ class DensityMatrixStepResult(simulator.StepResult):
             mixed state it must be correctly sized and positive semidefinite
             with trace one.
         """
-        density_matrix = density_matrix_utils.to_valid_density_matrix(
+        density_matrix = linalg.to_valid_density_matrix(
             density_matrix_repr,
             len(self._qubit_map),
             qid_shape=self._qid_shape,
