+++ cirq/sim/sparse_simulator.py
@@ -246,10 +246,10 @@ class Simulator(simulator.SimulatesSamples,
         num_qubits = len(qubits)
         qid_shape = protocols.qid_shape(qubits)
         qubit_map = {q: i for i, q in enumerate(qubits)}
-        state = wave_function.to_valid_state_vector(initial_state,
-                                                    num_qubits,
-                                                    qid_shape=qid_shape,
-                                                    dtype=self._dtype)
+        state = linalg.to_valid_state_vector(initial_state,
+                                             num_qubits,
+                                             qid_shape=qid_shape,
+                                             dtype=self._dtype)
         if len(circuit) == 0:
             yield SparseSimulatorStep(state, {}, qubit_map, self._dtype)
 
@@ -432,7 +432,7 @@ class SparseSimulatorStep(wave_function.StateVectorMixin,
         return self._simulator_state().state_vector
 
     def set_state_vector(self, state: 'cirq.STATE_VECTOR_LIKE'):
-        update_state = wave_function.to_valid_state_vector(
+        update_state = linalg.to_valid_state_vector(
             state,
             len(self.qubit_map),
             qid_shape=protocols.qid_shape(self, None),
